+++ Items/AsteroidShip.cpp
@@ -1006,7 +1006,7 @@ Weapon* AsteroidShip :: getPreviousWeapon() {
  * Increment the current weapon and mod by the number of weapons.
  */
 void AsteroidShip::nextWeapon() {
-   currentWeapon = (currentWeapon + 1) % weapons.size();
+   currentWeapon = (currentWeapon + 1) % (int) weapons.size();
 }
 
 /**
@@ -1014,7 +1014,7 @@ void AsteroidShip::nextWeapon() {
  */
 void AsteroidShip::prevWeapon() {
    // We add weaons.size() in the middle so that we don't do -1 % something.
-   currentWeapon = (currentWeapon + weapons.size() - 1) % weapons.size();
+   currentWeapon = (currentWeapon + (int) weapons.size() - 1) % (int) weapons.size();
 }
 
 // Return a reference to the list of weapons that the ship has.
@@ -1024,7 +1024,7 @@ std::vector<Weapon*> AsteroidShip :: getWeaponsList() {
 
 // Get the number of types of weapons the ship has. They're indexed 0 - (n-1)
 int AsteroidShip :: getNumWeapons() {
-   return weapons.size();
+   return (int) weapons.size();
 }
 
 /**
